+++ labkey-client-api/src/org/labkey/remoteapi/GuestCredentialsProvider.java
@@ -15,8 +15,8 @@
  */
 package org.labkey.remoteapi;
 
-import org.apache.http.client.methods.HttpUriRequest;
-import org.apache.http.client.protocol.HttpClientContext;
+import org.apache.hc.client5.http.classic.methods.HttpUriRequest;
+import org.apache.hc.client5.http.protocol.HttpClientContext;
 
 import java.net.URI;