+++ labkey-client-api/src/org/labkey/remoteapi/PostCommand.java
@@ -15,10 +15,10 @@
  */
 package org.labkey.remoteapi;
 
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.client.methods.HttpUriRequest;
-import org.apache.http.entity.ContentType;
-import org.apache.http.entity.StringEntity;
+import org.apache.hc.client5.http.classic.methods.HttpPost;
+import org.apache.hc.client5.http.classic.methods.HttpUriRequest;
+import org.apache.hc.core5.http.ContentType;
+import org.apache.hc.core5.http.io.entity.StringEntity;
 import org.json.simple.JSONObject;
 
 import java.net.URI;